Refactor `p4` into separate functional submodules
